@@ -13,7 +13,7 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3.0",
+		"php": ">=5.6.0",
 		"doctrine/collections": "~1.3"
 	},
 	"require-dev": {
@@ -22,7 +22,10 @@
 	"autoload": {
 		"psr-0": {
 			"Recurr": "src/"
-		}
+		},
+		"files": [
+			"src/Recurr/bootstrap.php"
+		]
 	},
 	"extra": {
 		"branch-alias": {
