@@ -15,11 +15,11 @@
 namespace Recurr;
 
 class Frequency {
-    const YEARLY   = 0;
-    const MONTHLY  = 1;
-    const WEEKLY   = 2;
-    const DAILY    = 3;
-    const HOURLY   = 4;
-    const MINUTELY = 5;
-    const SECONDLY = 6;
+    const YEARLY = 'YEARLY';
+    const MONTHLY = 'MONTHLY';
+    const WEEKLY = 'WEEKLY';
+    const DAILY = 'DAILY';
+    const HOURLY = 'HOURLY';
+    const MINUTELY = 'MINUTELY';
+    const SECONDLY = 'SECONDLY';
 }
