@@ -12,78 +12,50 @@
  */
 
 namespace Recurr;
+use DateInterval;
 use DateTimeImmutable;
+use InvalidArgumentException;
 
 /**
- * Defines a time of day.
+ * Defines a precise period of time.
  *
- * @see https://tools.ietf.org/html/rfc5545#section-3.3.12
+ * @see https://tools.ietf.org/html/rfc5545#section-3.3.9
  */
-final class Time
+final class Period
 {
-    const TZID = 'TZID';
-
-    const FLOATING_TIME_PATTERN = 'His';
-
-    const UTC_TIME_PATTERN = 'His\z';
 
     /**
-     * @var string The two-digit hour of the day (00-23).
+     * @var DateTime The start of the period.
      */
-    private $hour;
+    private $start;
 
     /**
-     * @var string The two-digit minute of the hour (00-59).
+     * @var DateTime The end of the period.
      */
-    private $minute;
+    private $end;
 
     /**
-     * @var string The two-digit second of the minute (00-60). A value of 60 represents a leap second.
+     * @var string Whether the format of the period is `period-explicit` or `period-start` (using a duration).
      */
-    private $second;
+    private $is_explicit;
 
-    /**
-     * @var string The time zone ID (TZID).
-     */
-    private $tzid;
-
-    /**
-     * Whether the time is in UTC.
-     *
-     * If this value is false, it is said to be "floating" and are not
-     * bound to any time zone in particular. They are used to represent
-     * the same hour, minute, and second value regardless of which
-     * time zone is currently being observed.
-     *
-     * @var bool
-     */
-    private $is_utc = false;
-
-    public function __construct($hour, $minute, $second, $tzid = null)
+    public function __construct(DateTime $start, DateTime $end = null, DateInterval $duration = null)
     {
-        if ('Z' === substr($time, -1, 1)) {
-            $this->is_utc = true;
-            $time = DateTimeImmutable::createFromFormat(self::UTC_TIME_PATTERN, $time);
-        } else {
-            $this->tzid = $tzid;
-            $time = DateTimeImmutable::createFromFormat(self::FLOATING_TIME_PATTERN, $time);
+        $this->start = $start;
+
+        if (null === $end && null === $duration) {
+            throw new InvalidArgumentException('The period must contain and end or a duration.');
+        } elseif (null === $end) {
+            $end = $start->toDateTimeImmutable()->add($duration);
+            $this->end = new DateTime()
         }
 
-        $this->hour = $time->format('H');
-        $this->minute = $time->format('i');
-        $this->second = $time->format('s');
+        $this->end = $end;
     }
 
-    public static function fromString($time)
+    public static function fromString($period)
     {
 
-        $date = DateTimeImmutable::createFromFormat(self::DATE_PATTERN, $date);
-
-        return static(
-            $date->format('Y'),
-            $date->format('m'),
-            $date->format('d')
-        )
     }
 
     /**
