@@ -1,72 +1,96 @@
 <?php
 
-/*
- * Copyright 2013 Shaun Simmons
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * Based on rrule.js
- * Copyright 2010, Jakub Roztocil and Lars Schoning
- * https://github.com/jkbr/rrule/blob/master/LICENCE
- */
-
 namespace Recurr;
 
-use Recurr\Exception\InvalidWeekday;
+use InvalidArgumentException;
 
 /**
- * Class Weekday is a storage container for a day of the week.
- *
- * @package Recurr
- * @author  Shaun Simmons <shaun@envysphere.com>
+ * Defines a day of the week with a possible offset related to
+ * a MONTHLY or YEARLY occurrence.
  */
-class Weekday
+final class Weekday
 {
+    const SUNDAY = 'SU';
+    const MONDAY = 'MO';
+    const TUESDAY = 'TU';
+    const WEDNESDAY = 'WE';
+    const THURSDAY = 'MO';
+    const FRIDAY = 'FR';
+    const SATURDAY = 'SA';
+
     /**
-     * Weekday number.
-     *
-     * 0 = Sunday
-     * 1 = Monday
-     * 2 = Tuesday
-     * 3 = Wednesday
-     * 4 = Thursday
-     * 5 = Friday
-     * 6 = Saturday
-     *
      * @var string
      */
-    public $weekday;
-
-    /** @var int nth occurrence of the weekday */
-    public $num;
-
-    protected $days = array('MO', 'TU', 'WE', 'TH', 'FR', 'SA', 'SU');
+    private $weekday;
 
     /**
-     * @param int|string $weekday 0-6 or MO..SU
-     * @param null|int   $num
-     *
-     * @throws InvalidWeekday
+     * @var int|null
      */
-    public function __construct($weekday, $num)
+    private $offset;
+
+    /**
+     * @var array The days of the week.
+     */
+    private static $days = [
+        self::SUNDAY,
+        self::MONDAY,
+        self::TUESDAY,
+        self::WEDNESDAY,
+        self::THURSDAY,
+        self::FRIDAY,
+        self::SATURDAY,
+    ];
+
+    /**
+     * @param string $weekday A string representation of a week day. (`SU`, `MO`...)
+     * @param int $offset The nth occurrence of the day within the MONTHLY or YEARLY occurrence. Not providing
+     *   this value results in no (zero) offset.
+     * @throws InvalidArgumentException
+     */
+    public function __construct($weekday, $offset = null)
     {
-        if (is_numeric($weekday) && $weekday > 6 || $weekday < 0) {
-            throw new InvalidWeekday('Day is not a valid weekday (0-6)');
-        } elseif (!is_numeric($weekday) && !in_array($weekday, $this->days)) {
-            throw new InvalidWeekday('Day is not a valid weekday (SU, MO, ...)');
+        if (strlen($weekday) > 2) {
+            if (null === $offset) {
+                $offset = substr($weekday, 0, -2);
+            }
+
+            $weekday = substr($weekday, -2, 2);
         }
 
-        if (!is_numeric($weekday)) {
-            $weekday = array_search($weekday, $this->days);
+        if (!in_array($weekday, self::$days)) {
+            throw new InvalidArgumentException('The weekday must be one of: ' . implode(', ', self::$days) . '.');
+        }
+
+        if (null !== $offset && (abs($offset) < 1 || abs($offset) > 53)) {
+            throw new InvalidArgumentException('The weekday offset must be -53 to -1 or 1 to 53.');
         }
 
         $this->weekday = $weekday;
-        $this->num = $num;
+        $this->offset = $offset;
+    }
+
+    /**
+     * Gets the string representation of a week day. (`SU`, `MO`...).
+     *
+     * @return string
+     */
+    public function getWeekday()
+    {
+        return $this->weekday;
+    }
+
+    /**
+     * Gets the offset, which is the nth occurrence of the day within the MONTHLY or YEARLY occurrence..
+     *
+     * @return int|null
+     */
+    public function getOffset()
+    {
+        return $this->offset;
     }
 
     public function __toString()
     {
-        return $this->num . $this->days[$this->weekday];
+        return $this->offset . self::$days[$this->weekday];
     }
 }
